+++ packages/frontend/src/components/Explorer/VisualizationCard/VisualizationCard.tsx
@@ -8,6 +8,7 @@ import useDashboardStorage from '../../../hooks/dashboard/useDashboardStorage';
 import { EChartSeries } from '../../../hooks/echarts/useEcharts';
 import { uploadGsheet } from '../../../hooks/gdrive/useGdrive';
 import { useExplore } from '../../../hooks/useExplore';
+import { useApp } from '../../../providers/AppProvider';
 import {
     ExplorerSection,
     useExplorerContext,
@@ -30,12 +31,15 @@ export type EchartsClickEvent = {
 
 const VisualizationCard: FC<{ projectUuid?: string }> = memo(
     ({ projectUuid: fallBackUUid }) => {
+        const { health } = useApp();
+
         // FEATURE FLAG: remove when chart config sidebar is complete
         const useSidebar = useFeatureFlagEnabled('sidebar-chart-config');
 
         const savedChart = useExplorerContext(
             (context) => context.state.savedChart,
         );
+
         const unsavedChartVersion = useExplorerContext(
             (context) => context.state.unsavedChartVersion,
         );
@@ -143,6 +147,11 @@ const VisualizationCard: FC<{ projectUuid?: string }> = memo(
             }
             throw new NotFoundError('no metric query defined');
         };
+
+        if (health.isLoading || !health.data) {
+            return null;
+        }
+
         return (
             <VisualizationProvider
                 initialChartConfig={unsavedChartVersion.chartConfig}
@@ -158,6 +167,7 @@ const VisualizationCard: FC<{ projectUuid?: string }> = memo(
                 onPivotDimensionsChange={setPivotFields}
                 columnOrder={unsavedChartVersion.tableConfig.columnOrder}
                 onSeriesContextMenu={onSeriesContextMenu}
+                pivotTableMaxColumnLimit={health.data.pivotTable.maxColumnLimit}
             >
                 <CollapsableCard
                     title="Charts"