+++ packages/frontend/src/hooks/pivotTable/pivotQueryResults.ts
@@ -9,8 +9,6 @@ import {
 import { isNumber } from 'lodash-es';
 import { Entries } from 'type-fest';
 
-const MAX_COLUMNS = 60;
-
 type PivotQueryResultsArgs = {
     pivotConfig: PivotConfig;
     metricQuery: Pick<
@@ -18,6 +16,9 @@ type PivotQueryResultsArgs = {
         'dimensions' | 'metrics' | 'tableCalculations' | 'additionalMetrics'
     >;
     rows: ResultRow[];
+    options: {
+        maxColumns: number;
+    };
 };
 
 type RecursiveRecord<T = unknown> = {
@@ -191,6 +192,7 @@ export const pivotQueryResults = ({
     pivotConfig,
     metricQuery,
     rows,
+    options,
 }: PivotQueryResultsArgs): PivotData => {
     if (rows.length === 0) {
         throw new Error('Cannot pivot results with no rows');
@@ -331,9 +333,9 @@ export const pivotQueryResults = ({
             ) {
                 columnCount++;
 
-                if (columnCount > MAX_COLUMNS) {
+                if (columnCount > options.maxColumns) {
                     throw new Error(
-                        `Cannot pivot results with more than ${MAX_COLUMNS} columns. Try adding a filter to limit your results.`,
+                        `Cannot pivot results with more than ${options.maxColumns} columns. Try adding a filter to limit your results.`,
                     );
                 }